Add support for multiple export versions
Google adds a "version" attribute to Classroom exports.
Classroom Export Viewer is currently designed only to work with version 1
exports, so a version check has been added in order to make it easier to
support future versions and to prevent incompatibilities with them
(by presenting the user with an error message if the version is not supported).