Remove Rollup and core-js
Rollup and core-js added a lot of complexity to the build process and
the result of transpilation, also meaning many more dependencies and
therefore a larger attack surface. Supporting IE9 is not important
enough to me to justify the additional maintenance; others are welcome
to transpile my code however they like. I plan on refactoring the code
to be more compatible at some point at which Babel will no longer be
needed (this will mean dropping support for some old browsers).